Add files via upload
With the growing migration of people to foreign countries in India, the Indian Government thought of conducting elections in an electronic way to make life easy for Non-Resident Indians via a Web Driven Application. It is not just limited to Non-residents, people who are busy with their routines can also vote via online Voting system, instead of standing in queue, which is the common method of voting. The main objective of this system is to minimize the errors of vote counting there by reducing congestion and to make use of all the voters. Voter has to pre-register the website before elections and can cast their vote to their party of interest. All the votes will be stored in the database, which will take very short period of time to declare the election results, which in turn results in mal-function or error of vote counting.